requirements.txt
paho-mqtt==2.*
requests==2.*
